+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/UploadBufferPool.java
@@ -1,9 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.storage.blob.specialized;
+package com.azure.storage.blob;
 
 import com.azure.core.util.logging.ClientLogger;
+import com.azure.storage.blob.specialized.BlockBlobClient;
 import com.azure.storage.common.Utility;
 import reactor.core.publisher.Flux;
