+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/specialized/BlockBlobAsyncClient.java
@@ -8,7 +8,6 @@ import com.azure.core.http.rest.Response;
 import com.azure.core.http.rest.SimpleResponse;
 import com.azure.core.implementation.util.FluxUtil;
 import com.azure.core.util.Context;
-import com.azure.core.util.logging.ClientLogger;
 import com.azure.storage.blob.BlobAsyncClient;
 import com.azure.storage.blob.implementation.AzureBlobStorageImpl;
 import com.azure.storage.blob.models.AccessTier;
@@ -21,34 +20,18 @@ import com.azure.storage.blob.models.BlockListType;
 import com.azure.storage.blob.models.BlockLookupList;
 import com.azure.storage.blob.models.CpkInfo;
 import com.azure.storage.blob.models.LeaseAccessConditions;
-import com.azure.storage.blob.models.ParallelTransferOptions;
 import com.azure.storage.blob.models.SourceModifiedAccessConditions;
 import com.azure.storage.common.Constants;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.UncheckedIOException;
 import java.net.URL;
 import java.nio.ByteBuffer;
-import java.nio.channels.AsynchronousFileChannel;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
-import java.util.ArrayList;
-import java.util.Base64;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
-import java.util.SortedMap;
-import java.util.TreeMap;
-import java.util.UUID;
-import java.util.stream.Collectors;
 
 import static com.azure.core.implementation.util.FluxUtil.withContext;
 import static com.azure.storage.blob.implementation.PostProcessor.postProcessResponse;
-import static java.nio.charset.StandardCharsets.UTF_8;
 
 /**
  * Client to a block blob. It may only be instantiated through a {@link SpecializedBlobClientBuilder} or via the method
@@ -67,11 +50,6 @@ import static java.nio.charset.StandardCharsets.UTF_8;
  */
 @ServiceClient(builder = SpecializedBlobClientBuilder.class, isAsync = true)
 public final class BlockBlobAsyncClient extends BlobAsyncClientBase {
-    private final ClientLogger logger = new ClientLogger(BlockBlobAsyncClient.class);
-
-    static final int BLOB_DEFAULT_UPLOAD_BLOCK_SIZE = 4 * Constants.MB;
-    static final int BLOB_MAX_UPLOAD_BLOCK_SIZE = 100 * Constants.MB;
-
     /**
      * Indicates the maximum number of bytes that can be sent in a call to upload.
      */
@@ -161,246 +139,6 @@ public final class BlockBlobAsyncClient extends BlobAsyncClientBase {
             null, data, length, null, metadata, tier, null, headers, accessConditions.getLeaseAccessConditions(),
             getCustomerProvidedKey(), accessConditions.getModifiedAccessConditions(), context))
             .map(rb -> new SimpleResponse<>(rb, new BlockBlobItem(rb.getDeserializedHeaders())));
-    }
-
-    /**
-     * Creates a new block blob, or updates the content of an existing block blob.
-     * <p>
-     * Updating an existing block blob overwrites any existing metadata on the blob. Partial updates are not supported
-     * with this method; the content of the existing blob is overwritten with the new content. To perform a partial
-     * update of a block blob's, use {@link BlockBlobAsyncClient#stageBlock(String, Flux, long) stageBlock} and {@link
-     * BlockBlobAsyncClient#commitBlockList(List)}. For more information, see the
-     * <a href="https://docs.microsoft.com/rest/api/storageservices/put-block">Azure Docs for Put Block</a> and the
-     * <a href="https://docs.microsoft.com/rest/api/storageservices/put-block-list">Azure Docs for Put Block List</a>.
-     * <p>
-     * The data passed need not support multiple subscriptions/be replayable as is required in other upload methods when
-     * retries are enabled, and the length of the data need not be known in advance. Therefore, this method should
-     * support uploading any arbitrary data source, including network streams. This behavior is possible because this
-     * method will perform some internal buffering as configured by the blockSize and numBuffers parameters, so while
-     * this method may offer additional convenience, it will not be as performant as other options, which should be
-     * preferred when possible.
-     * <p>
-     * Typically, the greater the number of buffers used, the greater the possible parallelism when transferring the
-     * data. Larger buffers means we will have to stage fewer blocks and therefore require fewer IO operations. The
-     * trade-offs between these values are context-dependent, so some experimentation may be required to optimize inputs
-     * for a given scenario.
-     *
-     * <p><strong>Code Samples</strong></p>
-     *
-     * {@codesnippet com.azure.storage.blob.specialized.BlockBlobAsyncClient.upload#Flux-ParallelTransferOptions}
-     *
-     * @param data The data to write to the blob. Unlike other upload methods, this method does not require that the
-     * {@code Flux} be replayable. In other words, it does not have to support multiple subscribers and is not expected
-     * to produce the same values across subscriptions.
-     * @param parallelTransferOptions {@link ParallelTransferOptions} used to configure buffered uploading.
-     * @return A reactive response containing the information of the uploaded block blob.
-     */
-    public Mono<BlockBlobItem> upload(Flux<ByteBuffer> data, ParallelTransferOptions parallelTransferOptions) {
-        return this.uploadWithResponse(data, parallelTransferOptions, null, null, null, null).flatMap(FluxUtil::toMono);
-    }
-
-    /**
-     * Creates a new block blob, or updates the content of an existing block blob. Updating an existing block blob
-     * overwrites any existing metadata on the blob. Partial updates are not supported with this method; the content of
-     * the existing blob is overwritten with the new content. To perform a partial update of a block blob's, use {@link
-     * BlockBlobAsyncClient#stageBlock(String, Flux, long) stageBlock} and
-     * {@link BlockBlobAsyncClient#commitBlockList(List)}, which this method uses internally. For more information,
-     * see the <a href="https://docs.microsoft.com/rest/api/storageservices/put-block">Azure
-     * Docs for Put Block</a> and the <a href="https://docs.microsoft.com/rest/api/storageservices/put-block-list">Azure
-     * Docs for Put Block List</a>.
-     * <p>
-     * The data passed need not support multiple subscriptions/be replayable as is required in other upload methods when
-     * retries are enabled, and the length of the data need not be known in advance. Therefore, this method should
-     * support uploading any arbitrary data source, including network streams. This behavior is possible because this
-     * method will perform some internal buffering as configured by the blockSize and numBuffers parameters, so while
-     * this method may offer additional convenience, it will not be as performant as other options, which should be
-     * preferred when possible.
-     * <p>
-     * Typically, the greater the number of buffers used, the greater the possible parallelism when transferring the
-     * data. Larger buffers means we will have to stage fewer blocks and therefore require fewer IO operations. The
-     * trade-offs between these values are context-dependent, so some experimentation may be required to optimize inputs
-     * for a given scenario.
-     *
-     * <p><strong>Code Samples</strong></p>
-     *
-     * {@codesnippet com.azure.storage.blob.specialized.BlockBlobAsyncClient.uploadWithResponse#Flux-ParallelTransferOptions-BlobHTTPHeaders-Map-AccessTier-BlobAccessConditions}
-     *
-     * @param data The data to write to the blob. Unlike other upload methods, this method does not require that the
-     * {@code Flux} be replayable. In other words, it does not have to support multiple subscribers and is not expected
-     * to produce the same values across subscriptions.
-     * @param parallelTransferOptions {@link ParallelTransferOptions} used to configure buffered uploading.
-     * @param headers {@link BlobHTTPHeaders}
-     * @param metadata Metadata to associate with the blob.
-     * @param tier {@link AccessTier} for the destination blob.
-     * @param accessConditions {@link BlobAccessConditions}
-     * @return A reactive response containing the information of the uploaded block blob.
-     */
-    public Mono<Response<BlockBlobItem>> uploadWithResponse(Flux<ByteBuffer> data,
-            ParallelTransferOptions parallelTransferOptions, BlobHTTPHeaders headers, Map<String, String> metadata,
-            AccessTier tier, BlobAccessConditions accessConditions) {
-        // TODO: Parallelism parameter? Or let Reactor handle it?
-        // TODO: Sample/api reference
-        Objects.requireNonNull(data, "data must not be null");
-        BlobAccessConditions accessConditionsFinal = accessConditions == null
-            ? new BlobAccessConditions() : accessConditions;
-        final ParallelTransferOptions finalParallelTransferOptions = parallelTransferOptions == null
-            ? new ParallelTransferOptions() : parallelTransferOptions;
-        int blockSize = finalParallelTransferOptions.getBlockSize();
-        int numBuffers = finalParallelTransferOptions.getNumBuffers();
-
-        // TODO: Progress reporting.
-        // See ProgressReporter for an explanation on why this lock is necessary and why we use AtomicLong.
-        /*AtomicLong totalProgress = new AtomicLong(0);
-        Lock progressLock = new ReentrantLock();*/
-
-        // Validation done in the constructor.
-        UploadBufferPool pool = new UploadBufferPool(numBuffers, blockSize);
-
-        /*
-        Break the source Flux into chunks that are <= chunk size. This makes filling the pooled buffers much easier
-        as we can guarantee we only need at most two buffers for any call to write (two in the case of one pool buffer
-        filling up with more data to write). We use flatMapSequential because we need to guarantee we preserve the
-        ordering of the buffers, but we don't really care if one is split before another.
-         */
-        Flux<ByteBuffer> chunkedSource = data.flatMapSequential(buffer -> {
-            if (buffer.remaining() <= blockSize) {
-                return Flux.just(buffer);
-            }
-            int numSplits = (int) Math.ceil(buffer.remaining() / (double) blockSize);
-            return Flux.range(0, numSplits)
-                .map(i -> {
-                    ByteBuffer duplicate = buffer.duplicate().asReadOnlyBuffer();
-                    duplicate.position(i * blockSize);
-                    duplicate.limit(Math.min(duplicate.limit(), (i + 1) * blockSize));
-                    return duplicate;
-                });
-        });
-
-        /*
-         Write to the pool and upload the output.
-         */
-        return chunkedSource.concatMap(pool::write)
-            .concatWith(Flux.defer(pool::flush))
-            .flatMapSequential(buffer -> {
-                // Report progress as necessary.
-                /*Flux<ByteBuffer> progressData = ProgressReporter.addParallelProgressReporting(Flux.just(buffer),
-                    optionsReal.progressReceiver(), progressLock, totalProgress);*/
-
-                final String blockId = Base64.getEncoder().encodeToString(
-                    UUID.randomUUID().toString().getBytes(UTF_8));
-
-                return this.stageBlockWithResponse(blockId, Flux.just(buffer), buffer.remaining(),
-                    accessConditionsFinal.getLeaseAccessConditions())
-                    // We only care about the stageBlock insofar as it was successful, but we need to collect the ids.
-                    .map(x -> {
-                        pool.returnBuffer(buffer);
-                        return blockId;
-                    }).flux();
-
-            }) // TODO: parallelism?
-            .collect(Collectors.toList())
-            .flatMap(ids ->
-                this.commitBlockListWithResponse(ids, headers, metadata, tier, accessConditions));
-
-    }
-
-    /**
-     * Creates a new block blob, or updates the content of an existing block blob, with the content of the specified
-     * file.
-     *
-     * <p><strong>Code Samples</strong></p>
-     *
-     * {@codesnippet com.azure.storage.blob.specialized.BlockBlobAsyncClient.uploadFromFile#String}
-     *
-     * @param filePath Path to the upload file
-     * @return An empty response
-     */
-    public Mono<Void> uploadFromFile(String filePath) {
-        return uploadFromFile(filePath, null, null, null, null, null);
-    }
-
-    /**
-     * Creates a new block blob, or updates the content of an existing block blob, with the content of the specified
-     * file.
-     *
-     * <p><strong>Code Samples</strong></p>
-     *
-     * {@codesnippet com.azure.storage.blob.specialized.BlockBlobAsyncClient.uploadFromFile#String-ParallelTransferOptions-BlobHTTPHeaders-Map-AccessTier-BlobAccessConditions}
-     *
-     * @param filePath Path to the upload file
-     * @param parallelTransferOptions {@link ParallelTransferOptions} to use to upload from file. Number of parallel
-     *        transfers parameter is ignored.
-     * @param headers {@link BlobHTTPHeaders}
-     * @param metadata Metadata to associate with the blob.
-     * @param tier {@link AccessTier} for the destination blob.
-     * @param accessConditions {@link BlobAccessConditions}
-     * @return An empty response
-     * @throws IllegalArgumentException If {@code blockSize} is less than 0 or greater than 100MB
-     * @throws UncheckedIOException If an I/O error occurs
-     */
-    // TODO (gapra) : Investigate if this is can be parallelized, and include the parallelTransfers parameter.
-    public Mono<Void> uploadFromFile(String filePath, ParallelTransferOptions parallelTransferOptions,
-        BlobHTTPHeaders headers, Map<String, String> metadata, AccessTier tier, BlobAccessConditions accessConditions) {
-        final ParallelTransferOptions finalParallelTransferOptions = parallelTransferOptions == null
-            ? new ParallelTransferOptions()
-            : parallelTransferOptions;
-
-        return Mono.using(() -> uploadFileResourceSupplier(filePath),
-            channel -> {
-                final SortedMap<Long, String> blockIds = new TreeMap<>();
-                return Flux.fromIterable(sliceFile(filePath, finalParallelTransferOptions.getBlockSize()))
-                    .doOnNext(chunk -> blockIds.put(chunk.getOffset(), getBlockID()))
-                    .flatMap(chunk -> {
-                        String blockId = blockIds.get(chunk.getOffset());
-                        return stageBlockWithResponse(blockId, FluxUtil.readFile(channel, chunk.getOffset(),
-                            chunk.getCount()), chunk.getCount(), null);
-                    })
-                    .then(Mono.defer(() -> commitBlockListWithResponse(
-                        new ArrayList<>(blockIds.values()), headers, metadata, tier, accessConditions)))
-                    .then()
-                    .doOnTerminate(() -> {
-                        try {
-                            channel.close();
-                        } catch (IOException e) {
-                            throw logger.logExceptionAsError(new UncheckedIOException(e));
-                        }
-                    });
-            }, this::uploadFileCleanup);
-    }
-
-
-    private AsynchronousFileChannel uploadFileResourceSupplier(String filePath) {
-        try {
-            return AsynchronousFileChannel.open(Paths.get(filePath), StandardOpenOption.READ);
-        } catch (IOException e) {
-            throw logger.logExceptionAsError(new UncheckedIOException(e));
-        }
-    }
-
-    private void uploadFileCleanup(AsynchronousFileChannel channel) {
-        try {
-            channel.close();
-        } catch (IOException e) {
-            throw logger.logExceptionAsError(new UncheckedIOException(e));
-        }
-    }
-
-    private String getBlockID() {
-        return Base64.getEncoder().encodeToString(UUID.randomUUID().toString().getBytes(StandardCharsets.UTF_8));
-    }
-
-    private List<BlobRange> sliceFile(String path, int blockSize) {
-        File file = new File(path);
-        assert file.exists();
-        List<BlobRange> ranges = new ArrayList<>();
-        for (long pos = 0; pos < file.length(); pos += blockSize) {
-            long count = blockSize;
-            if (pos + count > file.length()) {
-                count = file.length() - pos;
-            }
-            ranges.add(new BlobRange(pos, count));
-        }
-        return ranges;
     }
 
     /**
