+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/specialized/SpecializedBlobClientBuilder.java
@@ -73,9 +73,9 @@ public final class SpecializedBlobClientBuilder extends BaseBlobClientBuilder<Sp
 
     /**
      * Creates a {@link BlockBlobClient} based on options set in the Builder. BlockBlobClients are used to perform
-     * generic upload operations such as {@link BlockBlobClient#uploadFromFile(String) upload from file} and block blob
-     * specific operations such as {@link BlockBlobClient#stageBlock(String, InputStream, long) stage block} and {@link
-     * BlockBlobClient#commitBlockList(List)}, only use this when the blob is known to be a block blob.
+     * generic upload operations such as {@link BlockBlobClient#upload(InputStream, long) upload from file} and block
+     * blob specific operations such as {@link BlockBlobClient#stageBlock(String, InputStream, long) stage block} and
+     * {@link BlockBlobClient#commitBlockList(List)}, only use this when the blob is known to be a block blob.
      *
      * @return a {@link BlockBlobClient} created from the configurations in this builder.
      * @throws NullPointerException If {@code endpoint}, {@code containerName}, or {@code blobName} is {@code null}.
@@ -86,7 +86,7 @@ public final class SpecializedBlobClientBuilder extends BaseBlobClientBuilder<Sp
 
     /**
      * Creates a {@link BlockBlobAsyncClient} based on options set in the Builder. BlockBlobAsyncClients are used to
-     * perform generic upload operations such as {@link BlockBlobAsyncClient#uploadFromFile(String) upload from file}
+     * perform generic upload operations such as {@link BlockBlobAsyncClient#upload(Flux, long) upload from file}
      * and block blob specific operations such as {@link BlockBlobAsyncClient#stageBlockWithResponse(String, Flux, long,
      * LeaseAccessConditions) stage block} and {@link BlockBlobAsyncClient#commitBlockList(List) commit block list},
      * only use this when the blob is known to be a block blob.
